Moodle header changes
Moodle header color and title is set. But tabbarnav and navigation header can not be integrateg. Reason of this is that i can not access navigationOptions of a moodleDersDetay class because in the file, there no class named  moodleDersDetay. So navigation options of moodleDersDetay is given in the app.tsx and couldnt modified throughout app. This need to be discusssed.